bot/requirements.txt
python-telegram-bot[job-queue]==21.9
docker==7.1.0
pyyaml==6.0.2
psutil==6.1.1
croniter==2.0.7
wakeonlan==3.1.0
